promoteToSharedGreedy: do not intersect schedule with active points
A partial schedule produced by a call to partialSchedule already has its
domain intersected with the parent filter nodes.  There is no need to
separately compute the intersection of parent filters using
activeDomainPoints to include it in the schedule.  Drop the unnecessary
intersection and remove the variable that became unused.