promoteToSharedBelow: take into account the mapping to blocks
Currenly, promotion to shared memory is only performed below the loops
mapped to blocks.  Thus, tensor reference groups implictly account for
blocks.  The scope of mapping in the tree will be changed in upcoming
commits.  Explicitly include the block mapping into the partial schedule
within which the tensor reference groups are computed.