promoteToSharedAtDepth: ignore scopes below thread mapping
Band nodes at the given depth may be above or below the thread mapping.
In the latter case, promoteToSharedBelow would throw an IncorrectScope
exception because promotion to shared below a thread mapping node is
forbidden and is generally nonsensical.  In a branching tree, there is
no guarantee that nodes at the given schedule depth are all above or
below the thread mapping.  Simply ignore the subtrees that have a
thread-mapping ancestor node when performing shared memory promotion at
a given depth.

Leaving promoteToSharedBelow as is (throwing an exception) because it
may end up being reused by different callers.  Arguably,
promoteToSharedAtDepth is a higher-level API and should not throw if
only one subtree is problematic.