support promotion to shared memory at depth 0
The original implementation of shared memory promotion did not support
promotion to shared below a zero-dimensional outer band.  This was due
to bandsContainingScheduleDepth not supporting the depth 0, and is fixed
in a5a88f8 (bandsContainingScheduleDepth: support depth 0).  Drop the
check and the NYI exception from promoteToSharedAtDepth.  Drop the check
for the number of band members being non-zero in
makeWithOuterBlockInnerThreadStrategy.